requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
feedparser==6.0.11
html5lib==1.1
idna==3.7
lxml==5.2.1
PyPDF2==3.0.1
PyYAML==6.0.1
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tenacity==8.2.3
urllib3==2.2.1
webencodings==0.5.1
listparser
PyCryptodome
